Extract MembershipFormModal
